Refactoring and bug fix
